+++ sdk/paymentRequest/paymentRequest.js
@@ -121,7 +121,7 @@ exports.query = async function ({ centerId, limit, after, before, sort, status, 
      * @param limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
      * @param after [string, default null] date filter for objects created/due only after specified date. ex: '2020-03-10'
      * @param before [string, default null] date filter for objects created/due only before specified date. ex: '2020-03-10'
-     * @param sort[string, default '-created']: sort order considered in response.Valid options are '-created' or '-due'.
+     * @param sort [string, default '-created']: sort order considered in response.Valid options are '-created' or '-due'.
      * @param status [string, default null]: filter for status of retrieved objects.ex: 'success' or 'failed'
      * @param type [string, default null]: payment type , inferred from the payment parameter if it is not a dictionary.ex: 'transfer', 'boleto-payment'
      * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']