+++ sdk/webhook/webhook.js
@@ -8,12 +8,12 @@ class Webhook extends Resource {
      * Webhook subscription object
      *
      * @description A Webhook is used to subscribe to notification events on a user-selected endpoint.
-     * Currently available services for subscription are transfer, boleto, boleto-payment,
-     * and utility-payment
+     * Currently available services for subscription are transfer, invoice, deposit, brcode-payment,
+     * boleto, boleto-holmes, boleto-payment and utility-payment.
      *
      * Parameters (required):
      * @param url [string]: Url that will be notified when an event occurs.
-     * @param subscriptions [list of strings]: list of any non-empty combination of the available services. ex: ['transfer', 'boleto-payment']
+     * @param subscriptions [list of strings]: list of any non-empty combination of the available services. ex: ['transfer', 'deposit']
      * 
      * Attributes:
      * @param id [string, default null]: unique id returned when the webhook is created. ex: '5656565656565656'